requirements.txt
joblib==1.0.0
numpy==1.19.4
Pillow==8.0.1
pkg-resources
scikit-learn==0.23.2
scipy==1.5.4
sklearn
threadpoolctl==2.1.0
